# packages/server/src/app.ts
import express from 'express';import cors from 'cors';import helmet from 'helmet';import { z } from 'zod';import { existsSync } from 'node:fs';import { resolve } from 'node:path';import type { RequestHandler } from 'express';import type { Config } from './config.js';import type { Store } from './db.js';import type { AuthService } from './auth.js';import type { SessionService } from './service.js';
export function createApp(store:Store,auth:AuthService,sessions:SessionService,config:Config){const app=express();app.set('trust proxy',false);app.use(helmet({contentSecurityPolicy:false}));app.use(cors({origin:(origin,done)=>done(null,!origin||config.corsOrigins.includes(origin))}));app.use(express.json({limit:'1mb'}));
 app.get('/health',(_q,r)=>r.json({status:'ok',product:'codex',sessions_dir:config.codexSessionsDir,time:new Date().toISOString()}));
 app.post('/api/pair/code',(q,r)=>isLoopback(q.socket.remoteAddress)?r.status(201).json(auth.createPairCode()):r.status(403).json({error:'pair codes can only be generated locally'}));
 app.post('/api/pair/request',(q,r)=>{const x=z.object({code:z.string().regex(/^\d{6}$/),device_name:z.string().min(1).max(80)}).safeParse(q.body);if(!x.success)return r.status(400).json({error:'invalid request'});const t=auth.pair(x.data.code,x.data.device_name);return t?r.status(201).json(t):r.status(401).json({error:'invalid or expired code'})});
 app.post('/api/auth/refresh',(q,r)=>{const x=z.object({device_id:z.string().uuid(),refresh_token:z.string().min(20)}).safeParse(q.body);if(!x.success)return r.status(400).json({error:'invalid request'});const t=auth.refresh(x.data.device_id,x.data.refresh_token);return t?r.json(t):r.status(401).json({error:'invalid refresh token'})});
 const guard:RequestHandler=(q,r,n)=>{try{const h=q.headers.authorization;if(!h?.startsWith('Bearer '))throw 0;(q as any).device=auth.verify(h.slice(7));n()}catch{return r.status(401).json({error:'unauthorized'})}};app.use('/api',guard);
 app.get('/api/sessions',async(_q,r)=>r.json(await sessions.list()));
 const runtime=z.object({model:z.string().min(1).max(160).optional(),effort:z.enum(['none','minimal','low','medium','high','xhigh']).optional(),approvalPolicy:z.enum(['untrusted','on-failure','on-request','never']).optional(),sandbox:z.enum(['read-only','workspace-write','danger-full-access']).optional()}).strict();
 const userInput=z.discriminatedUnion('type',[z.object({type:z.literal('text'),text:z.string().min(1).max(200000)}).strict(),z.object({type:z.literal('image'),url:z.string().regex(/^data:image\/(?:png|jpeg|webp|gif);base64,/).max(8_000_000),name:z.string().max(255).optional()}).strict(),z.object({type:z.literal('localImage'),path:z.string().min(1).max(4096),name:z.string().max(255).optional()}).strict(),z.object({type:z.literal('skill'),name:z.string().min(1).max(160),path:z.string().min(1).max(4096)}).strict(),z.object({type:z.literal('mention'),name:z.string().min(1).max(255),path:z.string().min(1).max(4096)}).strict()]);
 app.post('/api/sessions',async(q,r)=>{const x=z.object({cwd:z.string().min(1).max(4096),runtime:runtime.optional()}).strict().safeParse(q.body);if(!x.success)return r.status(400).json({error:'invalid request'});try{return r.status(201).json(await sessions.create(x.data.cwd,x.data.runtime))}catch(e:any){return r.status(e.status||500).json({error:e.message})}});
 app.post('/api/sessions/refresh',async(_q,r)=>r.json(await sessions.refresh()));
 app.get('/api/sessions/:id',async(q,r)=>{const d=await sessions.detail(q.params.id);return d?r.json(d):r.status(404).json({error:'not found'})});
 app.patch('/api/sessions/:id',async(q,r)=>{const x=z.object({title:z.string().min(1).max(120).optional(),status:z.enum(['active','paused','archived']).optional(),pinned:z.boolean().optional()}).strict().refine(v=>Object.keys(v).length>0).safeParse(q.body);if(!x.success)return r.status(400).json({error:'invalid request'});const updated=await sessions.update(q.params.id,x.data);return updated?r.json(updated):r.status(404).json({error:'not found'})});
 app.delete('/api/sessions/:id',async(q,r)=>(await sessions.removeOverlay(q.params.id))?r.status(204).end():r.status(404).json({error:'not found'}));
 app.post('/api/sessions/:id/messages',async(q,r)=>{const x=z.object({role:z.literal('user').default('user'),content:z.string().min(1).max(200000).optional(),input:z.array(userInput).min(1).max(20).optional(),runtime:runtime.optional(),client_id:z.string().min(1).max(100).optional()}).strict().refine(v=>Boolean(v.content||v.input?.length)).safeParse(q.body);if(!x.success)return r.status(400).json({error:'invalid request'});try{return r.status(202).json(await sessions.message(q.params.id,x.data.input??[{type:'text',text:x.data.content!}],x.data.runtime,x.data.client_id))}catch(e:any){return r.status(e.status||500).json({error:e.message})}});
 app.post('/api/sessions/:id/cancel',async(q,r)=>{const x=z.object({}).strict().safeParse(q.body??{});if(!x.success)return r.status(400).json({error:'invalid request'});try{return r.status(202).json(await sessions.cancel(q.params.id))}catch(e:any){return r.status(e.status||500).json({error:e.message})}});
 app.get('/api/capabilities/models',async(_q,r)=>{try{return r.json(await sessions.models())}catch(e:any){return r.status(e.status||503).json({error:e.message})}});
 app.get('/api/capabilities/skills',async(q,r)=>{const x=z.object({cwd:z.string().min(1).max(4096)}).safeParse(q.query);if(!x.success)return r.status(400).json({error:'invalid query'});try{return r.json(await sessions.skills(x.data.cwd))}catch(e:any){return r.status(e.status||503).json({error:e.message})}});
 app.get('/api/capabilities/apps',async(q,r)=>{const x=z.object({thread_id:z.string().min(1).optional()}).safeParse(q.query);if(!x.success)return r.status(400).json({error:'invalid query'});try{return r.json(await sessions.apps(x.data.thread_id))}catch(e:any){return r.status(e.status||503).json({error:e.message})}});
 app.get('/api/capabilities/defaults',async(_q,r)=>{try{return r.json(await sessions.defaults())}catch(e:any){return r.status(e.status||503).json({error:e.message})}});
 app.get('/api/capabilities/cwd-roots',(_q,r)=>r.json(config.codexCwdAllowlist));
 app.get('/api/approvals',(q,r)=>{const x=z.object({session_id:z.string().min(1).optional()}).safeParse(q.query);return x.success?r.json(sessions.approvals(x.data.session_id)):r.status(400).json({error:'invalid query'})});
 app.get('/api/approvals/:id',(q,r)=>{const approval=sessions.approval(q.params.id);return approval?r.json(approval):r.status(404).json({error:'not found'})});
 app.post('/api/approvals/:id/decision',(q,r)=>{const x=z.object({decision:z.enum(['accept','decline','cancel']),answers:z.record(z.string().min(1),z.array(z.string()).min(1)).optional()}).strict().safeParse(q.body);if(!x.success)return r.status(400).json({error:'invalid request'});try{return r.json(sessions.decide(q.params.id,x.data.decision,x.data.answers))}catch(e:any){return r.status(e.status||500).json({error:e.message})}});
 app.get('/api/sync',(q,r)=>{const x=z.coerce.number().int().min(0).safeParse(q.query.cursor??0);if(!x.success)return r.status(400).json({error:'cursor must be a non-negative integer'});const events=store.eventsAfter(x.data);return r.json({cursor:events.at(-1)?.seq??x.data,events})});
 app.get('/api/files/search',async(q,r)=>{const x=z.object({query:z.string().min(1).max(200),roots:z.string().min(1).max(4096)}).safeParse(q.query);if(!x.success)return r.status(400).json({error:'invalid query'});try{return r.json(await sessions.fileSearch(x.data.query,x.data.roots.split(',').map(s=>s.trim()).filter(Boolean)))}catch(e:any){return r.status(e.status||503).json({error:e.message})}});
 app.get('/api/files/list',async(q,r)=>{const x=z.object({path:z.string().min(1).max(4096)}).safeParse(q.query);if(!x.success)return r.status(400).json({error:'invalid query'});try{return r.json(await sessions.readDirectory(x.data.path))}catch(e:any){return r.status(e.status||503).json({error:e.message})}});
 const webDist=resolve(process.cwd(),'packages/web/dist');if(existsSync(webDist)){app.use(express.static(webDist));app.get('*',(q,r,n)=>q.path.startsWith('/api/')||q.path==='/ws'?n():r.sendFile(resolve(webDist,'index.html')))}
 app.use((e:any,_q:any,r:any,_n:any)=>r.status(500).json({error:e.message||'internal error'}));return app}
function isLoopback(address:string|undefined){return address==='127.0.0.1'||address==='::1'||address?.startsWith('::ffff:127.')===true}